requirements.txt
alembic==1.13.2
aniso8601==9.0.1
astroid==3.3.3
asttokens==2.4.1
awscli==1.34.15
black==24.8.0
blinker==1.8.2
boto3==1.35.14
botocore==1.35.15
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.32.1
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
docker==7.1.0
docutils==0.16
efficientnet_pytorch==0.7.1
executing==2.1.0
filelock==3.16.0
fire==0.6.0
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.8
importlib_metadata==8.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mlflow==2.16.0
mlflow-skinny==2.16.0
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
plantuml==0.3.0
platformdirs==4.3.2
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==5.28.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.1
Pygments==2.18.0
pylint==3.3.0
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.34
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.2
termcolor==2.4.0
threadpoolctl==3.5.0
tomlkit==0.13.2
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchinfo==1.8.0
torchvision==0.19.0+cu124
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.5
wcwidth==0.2.13
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
zipp==3.20.1
